[move/#220] 스낵바 컴포넌트 위치 수정
